+++ src/libs/ascent/ascent.cpp
@@ -17,13 +17,19 @@
 
 #include <ascent_empty_runtime.hpp>
 #include <ascent_flow_runtime.hpp>
+#include <ascent_logging.hpp>
+#include <ascent_logging_old.hpp>
 #include <runtimes/ascent_main_runtime.hpp>
 #include <utils/ascent_string_utils.hpp>
 #include <flow.hpp>
 
+#include <conduit_fmt/conduit_fmt.h>
+
 #if defined(ASCENT_VTKH_ENABLED)
     #include <vtkh/vtkh.hpp>
 #endif
+
+
 
 #ifdef ASCENT_MPI_ENABLED
 #include <mpi.h>
@@ -37,12 +43,60 @@ namespace ascent
 {
 
 //-----------------------------------------------------------------------------
+namespace detail
+{
+
+//-----------------------------------------------------------------------------
+int
+ParRank(int comm_id)
+{
+    int rank = 0;
+
+#if defined(ASCENT_MPI_ENABLED)
+    if(comm_id == -1)
+    {
+      // do nothing, an error will be thrown later
+      // so we can respect the exception handling
+      return 0;
+    }
+    MPI_Comm mpi_comm = MPI_Comm_f2c(comm_id);
+    MPI_Comm_rank(mpi_comm, &rank);
+#endif
+
+    return rank;
+}
+
+//-----------------------------------------------------------------------------
+int
+ParSize(int comm_id)
+{
+int comm_size=1;
+
+#if defined(ASCENT_MPI_ENABLED)
+    if(comm_id == -1)
+    {
+      // do nothing, an error will be thrown later
+      // so we can respect the exception handling
+      return 1;
+    }
+    MPI_Comm mpi_comm = MPI_Comm_f2c(comm_id);
+    MPI_Comm_size(mpi_comm, &comm_size);
+#endif
+
+    return comm_size;
+}
+
+//-----------------------------------------------------------------------------
 void
 quiet_handler(const std::string &,
               const std::string &,
               int )
 {
 }
+
+}
+
+
 
 //-----------------------------------------------------------------------------
 Ascent::Ascent()
@@ -187,6 +241,9 @@ Ascent::open(const conduit::Node &options)
           comm_id = options["mpi_comm"].to_int32();
         }
 
+        int par_rank = detail::ParRank(comm_id);
+        int par_size = detail::ParSize(comm_id);
+
         CheckForSettingsFile(opts_file,
                              processed_opts,
                              true,
@@ -201,20 +258,183 @@ Ascent::open(const conduit::Node &options)
           m_options["mpi_comm"] = options["mpi_comm"];
         }
 
-        if(m_options.has_path("messages") &&
-           m_options["messages"].dtype().is_string() )
+        Node echo_opts;
+        echo_opts["echo_threshold"] = "info";
+        echo_opts["ranks"] = "root";
+
+        // messages echoed to std out
+        if(m_options.has_path("messages"))
         {
-            std::string msgs_opt = m_options["messages"].as_string();
-            if( msgs_opt == "verbose")
+            if(m_options["messages"].dtype().is_string())
             {
-                m_verbose_msgs = true;
+                std::string msgs_opt = m_options["messages"].as_string();
+                if( msgs_opt == "verbose")
+                {
+                    m_verbose_msgs = true;
+                    echo_opts["echo_threshold"] = "all";
+                }
+                else if(msgs_opt == "quiet")
+                {
+                    m_verbose_msgs = false;
+                    echo_opts["echo_threshold"] = "error";
+                }
             }
-            else if(msgs_opt == "quiet")
+            else if(m_options["messages"].dtype().is_object())
             {
-                m_verbose_msgs = false;
+                const Node &msg_ops = m_options["messages"];
+                echo_opts.update(msg_ops);
+                if(msg_ops.has_child("echo_threshold") &&
+                   msg_ops["echo_threshold"].dtype().is_string() ) 
+                {
+                    std::string echo_thresh = msg_ops["echo_threshold"].as_string();
+                    if(echo_thresh == "none" || 
+                       echo_thresh == "error" ||
+                       echo_thresh == "warn" )
+                    {
+                        m_verbose_msgs = false;
+                    }
+                    else
+                    {
+                        m_verbose_msgs = true;
+                    }
+                }
+            }
+        }
+        
+        ascent::Logger &logger = ascent::Logger::instance();
+
+        // setup echo
+        logger.set_echo_threshold(echo_opts["echo_threshold"].as_string());
+
+        // controls for mpi ranks
+        // if ranks == "root"
+        //   rank 0 is what is specified, echo_threshold = "none" for all others
+        // if ranks == "all"
+        //   echo_threshold = specified option used for all ranks (alreay handled above)
+        // if ranks == [list of ints] (also accepts single int)
+        //   echo_threshold = specified option used for ranks in the list
+
+        if(echo_opts["ranks"].dtype().is_number()) // list of ints case
+        {
+            int64_accessor ranks_list = echo_opts["ranks"].value();
+            bool active = false;
+            for(index_t i=0; i < ranks_list.number_of_elements(); i++)
+            {
+                if(par_rank == ranks_list[i] )
+                {
+                    active = true;
+                }
+            }
+
+            if(!active)
+            {
+                logger.set_echo_threshold("none");
+            }
+        }
+        else // string options case
+        {
+            std::string log_ranks_str = echo_opts["ranks"].as_string();
+
+            if(log_ranks_str == "root")
+            {
+                if(par_rank != 0)
+                {
+                    logger.set_echo_threshold("none");
+                }
             }
         }
 
+        // logging options
+        //
+        //   logging: true
+        //
+        //   logging:
+        //      file_pattern: zzzz
+        //      log_threshold:  info
+        //
+
+        Node logging_opts;
+        logging_opts["enabled"] = 0;
+        logging_opts["ranks"]   = "all";
+#if defined(ASCENT_MPI_ENABLED)
+        logging_opts["file_pattern"]  = "ascent_log_output_rank_{rank:05d}.yaml";
+#else
+        logging_opts["file_pattern"]  = "ascent_log_output.yaml";
+#endif
+        logging_opts["log_threshold"] = "debug";
+
+        if(m_options.has_path("logging"))
+        {
+            if(m_options["logging"].dtype().is_string() &&
+               m_options["logging"].as_string() == "true")
+            {
+                logging_opts["enabled"] = 1;
+            }
+            else if(m_options["logging"].dtype().is_object())
+            {
+                logging_opts["enabled"] = 1;
+                // pull over options
+                logging_opts.update(m_options["logging"]);
+            }
+        }
+
+        // controls for mpi ranks
+        // if ranks == "root"
+        //   open log on rank 0, do not open on all others
+        // if ranks == "all"
+        //   open log on all ranks
+        // if ranks == [list of ints] (also accepts single int)
+        //   open log on ranks specified in the list
+
+        if(logging_opts["ranks"].dtype().is_number()) // list of ints case
+        {
+            int64_accessor ranks_list = logging_opts["ranks"].value();
+            bool active = false;
+            for(index_t i=0; i < ranks_list.number_of_elements(); i++)
+            {
+                if(par_rank == ranks_list[i] )
+                {
+                    active = true;
+                }
+            }
+
+            if(!active)
+            {
+                logging_opts["enabled"] = 0;
+            }
+        }
+        else // string options case
+        {
+            std::string log_ranks_str = logging_opts["ranks"].as_string();
+
+            if(log_ranks_str == "root")
+            {
+                if(par_rank != 0)
+                {
+                    logging_opts["enabled"] = 0;
+                }
+            }
+            // all already supported if logging is enabled
+        }
+
+
+        if(logging_opts["enabled"].to_int() == 1)
+        {
+            logger.set_log_threshold(logging_opts["log_threshold"].as_string());
+            std::string file_pattern = logging_opts["file_pattern"].as_string();
+        #if defined(ASCENT_MPI_ENABLED)
+            ASCENT_LOG_OPEN_RANK( file_pattern, par_rank ) // mpi par
+            ASCENT_LOG_DEBUG(conduit_fmt::format("mpi info: rank={}, size={}",
+                                                  par_rank,
+                                                  par_size));
+        #else
+            ASCENT_LOG_OPEN( file_pattern ) // serial
+            ASCENT_LOG_DEBUG("mpi not enabled");
+        #endif
+        }
+
+
+        // exception controls
         if(m_options.has_path("exceptions") &&
            m_options["exceptions"].dtype().is_string() )
         {
@@ -300,7 +520,7 @@ Ascent::open(const conduit::Node &options)
         // make sure to do this after.
         if(!m_verbose_msgs)
         {
-            conduit::utils::set_info_handler(quiet_handler);
+            conduit::utils::set_info_handler(detail::quiet_handler);
         }
 
         set_status("Ascent::open completed");
@@ -549,6 +769,7 @@ Ascent::close()
         }
 
          set_status("Ascent::close completed");
+         ASCENT_LOG_CLOSE();
     }
     catch(conduit::Error &e)
     {
@@ -574,6 +795,7 @@ Ascent::close()
                         << e.message() << std::endl;
             }
         }
+        ASCENT_LOG_CLOSE();
     }
 }
 
@@ -585,6 +807,7 @@ Ascent::set_status(const std::string &msg)
     std::ostringstream oss;
     oss << msg << " at " << timestamp();
     m_status["message"] = oss.str();
+    ASCENT_LOG_DEBUG(msg);
 }
 
 //---------------------------------------------------------------------------//
@@ -597,6 +820,7 @@ Ascent::set_status(const std::string &msg,
     oss << msg << " at " << timestamp();
     m_status["message"] = oss.str();
     m_status["details"] = details;
+    ASCENT_LOG_DEBUG(msg + " " + details);
 }
 
 //---------------------------------------------------------------------------//