+++ src/libs/ascent/runtimes/ascent_transmogrifier.cpp
@@ -17,6 +17,7 @@
 #include "ascent_mfem_data_adapter.hpp"
 #endif
 #include "ascent_logging.hpp"
+#include "ascent_logging_old.hpp"
 #include <conduit_blueprint.hpp>
 #include <algorithm>