+++ CHANGELOG.md
@@ -12,9 +12,11 @@ and this project aspires to adhere to [Semantic Versioning](https://semver.org/s
 - raja@2024.02.1
 - umpire@2024.02.1
 - camp@2024.02.1
-- kokkos@3.7.02
+- kokkos@4.4.1
+- mfem@4.7
 
 ### Added
+- Added a new unified logging infrastructure.
 - Added support for unstructured topologies with mixed elements types (for example, hexs and tets).
 - Added support for `pyramid` and `wedge` elements.
 - Added `sphere`, `cylinder`, `box`, and `plane` options to the slice filter.
@@ -29,6 +31,8 @@ and this project aspires to adhere to [Semantic Versioning](https://semver.org/s
 ### Fixed
 - Resolved a few cases where MPI_COMM_WORLD was used instead instead of the selected MPI communicator.
 - Resolved a bug where a sharing a coordset between multiple polytopal topologies would corrupt mesh processing.
+- Fixed a bug with Cinema resource output that could lead to corrupted html results.
+- Fixed a bug where controls for world and screen annotations where ignored in Cinema renders.
 
 ## [0.9.3] - Released 2024-05-11
 ### Preferred dependency versions for ascent@0.9.3